requirements.txt
python-dotenv
aiogram
fastapi
stripe
asyncio
pyngrok
uvicorn
sqlalchemy
aiosqlite